Documents/Zach Wilson DataEngineering DataExpert.io Bootcamp/AI_Engineering/fast-api-vibe-coding/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai>=1.12.0
python-multipart==0.0.6
python-dotenv==1.0.0
grpcio>=1.54.0
pymilvus==2.3.4
mmh3==4.0.1

